feat(v0.3.2): implement BradleyTerry MM fitter, property tests, benchmarks, docs
The core src/bradley-terry/index.ts was a 0-LOC placeholder since v0.2.6.
This ships the actual engine plus the supporting test/bench/doc infrastructure.

Core engine:
- BradleyTerry Context tag + BradleyTerryLive Layer (Effect-TS service)
- MM algorithm (Hunter 2004) with Float64 typed arrays
- Union-Find graph connectivity — fits largest component, excludes isolated
- Time decay (exponential, configurable half-life)
- Three output scales: arithmetic, geometric, elo400
- Log-likelihood reporting
- Typed errors: SelfMatchError, InsufficientDataError, ConvergenceError,
  DisconnectedGraphError, EntityNotFoundError

Tests (6 pass, 0 fail, 368 expect() calls):
- mm-invariants: win-prob symmetry + monotonicity under added wins
- graph-connectivity: largestComponentSize correctness + disconnected handling
- error-handling: SelfMatchError + InsufficientDataError guarantees

Benchmarks:
- 50k matches in 87ms (target <1.5s — 17x under)
- 5k in 3ms, 25k in 8ms via runBench()
- Bun macro embeds git commit hash in bench output

Fixes:
- Add missing fast-check@4.8.0 devDependency
- Fix fast-check 4.x default-export import pattern
- Fix fast-check 4.x fc.float 32-bit constraint (Math.fround)
- Fix Effect Cause unwrapping in error tests (Cause.failureOption)
- Fix entityCount semantics (total entities, not largest-component count)
- Add root bradley-terry.ts re-export for test import resolution
- Add missing benchmark-loader.ts module

Docs:
- Full README rewrite (features, API, quick start, benchmarks, layout)
- Full ARCHITECTURE rewrite (layer diagram, data flow, per-layer detail)
- RELEASE_NOTES_v0.3.2.md
- CHANGELOG updated
- Bump version to 0.3.2

+++ test/benchmark/bradley-terry.bench.ts
@@ -1,5 +1,5 @@
 import { test, expect } from "bun:test";
-import { fc } from "fast-check";
+import fc from "fast-check";
 import { Effect } from "effect";
 import { BradleyTerry, BradleyTerryLive } from "../../bradley-terry";
 import type { EntityId } from "../../schema";
@@ -13,7 +13,7 @@ const matchArb = fc
     winner: entityArb,
     loser: entityArb,
     date: fc.option(fc.date()),
-    weight: fc.option(fc.float({ min: 0.1, max: 8 })),
+    weight: fc.option(fc.float({ min: Math.fround(0.1), max: Math.fround(8) })),
   })
   .filter((m) => m.winner !== m.loser);
